+++ deep-links/DynamicLinksSandbox/android/app/BUCK
@@ -35,12 +35,12 @@ android_library(
 
 android_build_config(
     name = "build_config",
-    package = "com.dynamiclinkssandbox",
+    package = "ws4020.reactnative.sandbox.deeplink",
 )
 
 android_resource(
     name = "res",
-    package = "com.dynamiclinkssandbox",
+    package = "ws4020.reactnative.sandbox.deeplink",
     res = "src/main/res",
 )
