+++ deep-links/DynamicLinksSandbox/android/app/src/debug/ws4020/reactnative/sandbox/deeplink/ReactNativeFlipper.java
@@ -4,7 +4,7 @@
  * <p>This source code is licensed under the MIT license found in the LICENSE file in the root
  * directory of this source tree.
  */
-package com.dynamiclinkssandbox;
+package ws4020.reactnative.sandbox.deeplink;
 
 import android.content.Context;
 import com.facebook.flipper.android.AndroidFlipperClient;
