+++ deep-links/DynamicLinksSandbox/android/app/src/main/AndroidManifest.xml
@@ -1,4 +1,4 @@
-<manifest xmlns:android="http://schemas.android.com/apk/res/android" package="com.dynamiclinkssandbox">
+<manifest xmlns:android="http://schemas.android.com/apk/res/android" package="ws4020.reactnative.sandbox.deeplink">
 
   <uses-permission android:name="android.permission.INTERNET"/>
