+++ deep-links/DynamicLinksSandbox/android/app/src/main/java/ws4020/reactnative/sandbox/deeplink/MainActivity.java
@@ -1,4 +1,4 @@
-package com.dynamiclinkssandbox;
+package ws4020.reactnative.sandbox.deeplink;
 
 import android.os.Bundle;
