+++ deep-links/DynamicLinksSandbox/android/app/src/main/java/ws4020/reactnative/sandbox/deeplink/MainApplication.java
@@ -1,30 +1,22 @@
-package com.dynamiclinkssandbox;
+package ws4020.reactnative.sandbox.deeplink;
 
 import android.app.Application;
 import android.content.Context;
-import android.net.Uri;
 
 import com.facebook.react.PackageList;
 import com.facebook.react.ReactApplication;
 import com.facebook.react.ReactInstanceManager;
 import com.facebook.react.ReactNativeHost;
 import com.facebook.react.ReactPackage;
-import com.facebook.react.shell.MainReactPackage;
 import com.facebook.soloader.SoLoader;
-import com.dynamiclinkssandbox.generated.BasePackageList;
+import ws4020.reactnative.sandbox.deeplink.generated.BasePackageList;
 
-import org.unimodules.adapters.react.ReactAdapterPackage;
 import org.unimodules.adapters.react.ModuleRegistryAdapter;
 import org.unimodules.adapters.react.ReactModuleRegistryProvider;
-import org.unimodules.core.interfaces.Package;
-import org.unimodules.core.interfaces.SingletonModule;
-import expo.modules.constants.ConstantsPackage;
-import expo.modules.permissions.PermissionsPackage;
-import expo.modules.filesystem.FileSystemPackage;
+
 import expo.modules.updates.UpdatesController;
 
 import java.lang.reflect.InvocationTargetException;
-import java.util.Arrays;
 import java.util.List;
 import javax.annotation.Nullable;
 
@@ -102,7 +94,7 @@ public class MainApplication extends Application implements ReactApplication {
          We use reflection here to pick up the class that initializes Flipper,
         since Flipper library is not available in release mode
         */
-        Class<?> aClass = Class.forName("com.dynamiclinkssandbox.ReactNativeFlipper");
+        Class<?> aClass = Class.forName("ws4020.reactnative.sandbox.deeplink.ReactNativeFlipper");
         aClass
             .getMethod("initializeFlipper", Context.class, ReactInstanceManager.class)
             .invoke(null, context, reactInstanceManager);
